move to separate file with tests
